+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-12-04
+### Added
+- Allow specifying "api" as the channel to receive generated responses
+- Allow for passing the use case on API calls
+
+### Changed
+- Triggers and API now send single inbound messages instead of an array of inbound messages
+- Updated documentation
+
 ## [2.0.0] - 2025-07-02
 
 ### Added